requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
anykeystore==0.2
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
audiomentations==0.28.0
audioread==3.0.0
autopep8==2.0.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.0
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.0.3
colorama==0.4.4
comm==0.1.2
command-not-found==0.3
contourpy==1.0.7
cryptacular==1.6.2
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
datasets @ git+https://github.com/huggingface/datasets.git@a0a35c5fa84a8a7df656c1f5b0a7266126fa9b75
dbus-python==1.2.18
debugpy==1.6.6
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
dill==0.3.6
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
executing==1.2.0
fasteners==0.14.1
fastjsonschema==2.16.2
filelock==3.9.0
flatbuffers==23.1.21
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.3.0
future==0.18.2
gast==0.4.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googletrans==4.0.0rc1
greenlet==2.0.2
grpcio==1.51.3
h11==0.9.0
h2==3.2.0
h5py==3.8.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.20.2
httpx==0.13.3
huggingface-hub==0.12.1
hupper==1.11
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.6.4
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.1
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keras==2.11.0
keras-tuner==1.3.0
keyring==23.5.0
kiwisolver==1.4.4
kt-legacy==1.0.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.1
libclang==15.0.6.1
librosa==0.9.2
llvmlite==0.39.1
lockfile==0.12.2
loguru==0.6.0
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mistune==2.0.5
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.4
multidict==6.0.4
multiprocess==0.70.14
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
netifaces==0.11.0
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.0
olefile==0.46
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
PasteDeploy==3.0.1
pbkdf2==1.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==3.0.0
playsound==1.3.0
pooch==1.6.0
prometheus-client==0.16.0
prompt-toolkit==3.0.37
protobuf==3.12.4
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.21
pycups==2.0.1
pydot==1.4.2
pydub==0.25.1
pygame==2.2.0
Pygments==2.14.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyramid==2.0.1
pyramid-mailer==0.15.1
pyRFC3339==1.1
pyrsistent==0.19.3
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-json-logger==2.0.7
python3-openid==3.2.0
pytube==12.1.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
reportlab==3.6.8
repoze.sendmail==4.4.1
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rsa==4.9
sacremoses==0.0.53
scikit-learn==1.2.1
scipy==1.10.1
SecretStorage==3.3.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.4
soxr==0.3.3
SpeechRecognition==3.9.0
SQLAlchemy==1.4.46
stack-data==0.6.2
systemd-python==234
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.1
torchaudio==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transaction==3.0.1
transformers @ git+https://github.com/huggingface/transformers.git@4130e7036793ab2da08c1f3992affa1e56349ef8
translationstring==1.4
typing_extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.2.0
urllib3==1.26.5
usb-creator==0.3.7
velruse==1.1.1
venusian==3.0.0
wadllib==1.3.6
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
WebOb==1.8.7
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.5
wrapt==1.14.1
WTForms==3.0.1
wtforms-recaptcha==0.3.2
xdg==5
xkit==0.0.0
xxhash==3.2.0
yarl==1.8.2
zipp==1.0.0
zope.deprecation==4.4.0
zope.interface==5.5.2
zope.sqlalchemy==2.0
